@@ -16,7 +16,7 @@
 
 import { NgModule } from '@angular/core';
 import { ReplConsoleComponent } from './repl-console-component';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatIconModule } from '@angular/material/icon';
@@ -26,22 +26,16 @@ import { ReplResultDetailComponent } from './repl-result-detail-component';
 import { SharedModule } from '../shared/shared-module';
 
 
-@NgModule({
-  declarations: [
-    ReplConsoleComponent,
-    ReplResultDetailComponent,
-  ],
-  imports: [
-    CommonModule,
-    HttpClientModule,
-    MatFormFieldModule,
-    SharedModule,
-    MatInputModule,
-    MatIconModule,
-    MatButtonModule,
-  ],
-  exports: [
-    ReplConsoleComponent
-  ]
-})
+@NgModule({ declarations: [
+        ReplConsoleComponent,
+        ReplResultDetailComponent,
+    ],
+    exports: [
+        ReplConsoleComponent
+    ], imports: [CommonModule,
+        MatFormFieldModule,
+        SharedModule,
+        MatInputModule,
+        MatIconModule,
+        MatButtonModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class ReplConsoleModule { }
