@@ -15,17 +15,19 @@
  */
 
 import { TestBed } from '@angular/core/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 
 import { ReplApiService } from './repl-api-service';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('ReplApiService', () => {
   let service: ReplApiService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule]
-    });
+    imports: [],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+});
     service = TestBed.inject(ReplApiService);
   });
 
